Reverted WrappedActivity & ActivityView
